fix skins: store paint kit and seed as float, not uint32
items_game.txt defines attr 6 (paint kit) and attr 7 (seed) as
stored_as_integer=float_floor_to_integer. CS2 reads value_bytes as IEEE
float then floor()s to get the integer. We were storing raw uint32 bytes
(e.g. uint32(44) = 0x0000002C), which CS2 reads as float 5.6e-44,
floors to 0, and sets m_nFallbackPaintKit=0 — no skin rendered.

Fix: change attr 6 and 7 to AttributeType::Float so SetAttributeString
stores float(44.0f) = 0x42300000, which CS2 correctly reads as 44.

+++ csgo_gc/item_schema.cpp
@@ -115,8 +115,8 @@ ItemSchema::ItemSchema()
     // items_game.txt is unavailable (CS2 packs it in VPK, not a loose file).
     // items_game.txt overrides these via try_emplace after loading.
     static const std::pair<uint32_t, AttributeType> k_knownAttributes[] = {
-        { 6,   AttributeType::Uint32 }, // set item texture prefab (paint kit)
-        { 7,   AttributeType::Uint32 }, // set item texture seed
+        { 6,   AttributeType::Float  }, // set item texture prefab (paint kit) — stored_as_integer=float_floor_to_integer
+        { 7,   AttributeType::Float  }, // set item texture seed — stored_as_integer=float_floor_to_integer
         { 8,   AttributeType::Float  }, // set item texture wear
         { 80,  AttributeType::Uint32 }, // kill eater (StatTrak count)
         { 81,  AttributeType::Uint32 }, // kill eater score type
